Made challenge links on the chapters page clickable
This is part of issue #10. The Challenges component previously only received and showed a list of lesson IDs. With these changes it finds the respective challenge data to show the title and a link to the respective challenge content.

Note that the links don't work right now. Contentlayer is not producing the correct links we need. This can be handled in a separate PR. Maybe a good task for Juan.